Update Cypress tests
