+++ .github/workflows/ci.yml
@@ -7,7 +7,7 @@ jobs:
     name: Build and Run Jest Tests
     runs-on: ubuntu-latest
     steps:
-      - uses: actions/checkout@v2
+      - uses: actions/checkout@v4
       - uses: actions/setup-node@v2
       - name: Install Dependencies
         run: npm ci
@@ -33,8 +33,8 @@ jobs:
         containers: [1, 2, 3]
     steps:
       - name: Checkout
-        uses: actions/checkout@v2
-      - uses: cypress-io/github-action@v2
+        uses: actions/checkout@v4
+      - uses: cypress-io/github-action@v4
         with:
           start: npm start
           wait-on: 'http://localhost:8080'
@@ -56,16 +56,17 @@ jobs:
           # by default.
           CYPRESS_coverage: true
       - name: Upload coverage to Codecov
-        uses: codecov/codecov-action@v1
+        uses: codecov/codecov-action@v2
         with:
           flags: cypress
+          token: ${{ secrets.CODECOV_TOKEN }}
   s3-deploy:
     name: S3 Deploy
     needs:
       - build_test
     runs-on: ubuntu-latest
     steps:
-      - uses: actions/checkout@v2
+      - uses: actions/checkout@v4
       - uses: actions/setup-node@v2
       - name: Install Dependencies
         run: npm ci
