+++ cypress.config.ts
@@ -1,0 +1,20 @@
+import { defineConfig } from 'cypress'
+
+export default defineConfig({
+  video: false,
+  fixturesFolder: false,
+  projectId: 'u9ub4z',
+  defaultCommandTimeout: 8000,
+  env: {
+    coverage: false,
+  },
+  e2e: {
+    // We've imported your old cypress plugins here.
+    // You may want to clean this up later by importing these.
+    setupNodeEvents(on, config) {
+      return require('@cypress/code-coverage/task')(on, config)
+    },
+    baseUrl: 'http://localhost:8080',
+    specPattern: 'cypress/e2e/**/*.{js,jsx,ts,tsx}',
+  },
+})
