--- cypress.json
@@ -1,10 +1,0 @@
-{
-  "baseUrl": "http://localhost:8080",
-  "video": false,
-  "fixturesFolder": false,
-  "projectId": "u9ub4z",
-  "defaultCommandTimeout": 8000,
-  "env": {
-    "coverage": false
-  }
-}
