--- cypress/plugins/index.js
@@ -1,6 +1,0 @@
-module.exports = (on, config) => {
-  require("@cypress/code-coverage/task")(on, config);
-
-  // it is required that the config is returned
-  return config;
-};
